requirements.txt
Flask_Caching==1.8.0
PyMySQL==0.9.3
flask_restplus==0.13.0
Flask_Cors==3.0.8
Flask_DebugToolbar==0.11.0
Flask_Login==0.5.0
Flask==1.1.2
Flask_Bcrypt==0.7.1
environs==7.3.0
Flask_Static_Digest==0.1.3
loguru==0.4.1
SQLAlchemy==1.3.15
click==6.7
arrow==0.15.5
Flask_SQLAlchemy==2.4.1
Flask_Migrate==2.5.2
PyJWT==1.7.1
PyYAML==5.3.1
